2. Парковка машин [#853 #182301] final

+++ src/main/java/ru/job4j/lsp/Car.java
@@ -1,5 +1,5 @@
 package ru.job4j.lsp;
 
 public interface Car {
-    double getSize();
+    int getSize();
 }
